+++ contracts/game-of-chance/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2.1] - 2022-12-14
+### Changed
+- Partially restored `GOCState` from the `0.1.1` version.
+- Winner sets to `ActorId::zero()` after every game (re)start.
 
 ## [0.2.0] - 2022-11-12
 ### Added
@@ -22,7 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/gear-dapps/lottery/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/gear-dapps/lottery/compare/0.2.1...HEAD
+[0.2.1]: https://github.com/gear-dapps/lottery/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/gear-dapps/lottery/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/gear-dapps/lottery/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/gear-dapps/lottery/compare/60d5a8e...0.1.0